+++ saleor/static/dashboard-next/collections/queries.ts
@@ -89,7 +89,9 @@ export const collectionDetails = gql`
               id
               name
             }
-            thumbnailUrl
+            thumbnail {
+              url
+            }
           }
         }
         pageInfo {