+++ CHANGELOG.md
@@ -30,3 +30,4 @@ All notable, unreleased changes to this project will be documented in this file.
 - Remove Elasticsearch from `docker-compose.yml` - #3482 by @maarcingebala
 - Add error page if object was not found - #3463 by @dominik-zeglen
 - Fix simple product's inventory data saving bug - #3474 by @dominik-zeglen
+- Replace `thumbnailUrl` with `thumbnail { url }` - #3484 by @dominik-zeglen